Bluetooth 기기 초기값 데이터 전송 코드 리팩토링
